toolkit/content/tests/chrome/test_bug460942.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=460942
-->
<window title="Mozilla Bug 460942"
        onload="runTests()"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=460942"
     target="_blank">Mozilla Bug 460942</a>
  </body>

  <!-- test code goes here -->

  <richlistbox>
    <richlistitem id="item1">
        <label value="one"/>
        <box>
            <label value="two"/>
        </box>
    </richlistitem>
    <richlistitem id="item2"><description>one</description><description>two</description></richlistitem>
  </richlistbox>

  <script type="application/javascript">
  <![CDATA[
  /** Test for Bug 460942 **/
  function runTests() {
    is ($("item1").label, "one two");
    is ($("item2").label, "");
    SimpleTest.finish();
  }
  SimpleTest.waitForExplicitFinish();
  ]]>
  </script>

</window>
